auto-claude: subtask-1-2 - Extract match generation utilities to utils/matchGenerator.ts
