+++ .auto-claude-status
@@ -1,10 +1,10 @@
 {
   "active": true,
   "spec": "003-split-monolithic-app-tsx-into-smaller-components",
-  "state": "planning",
+  "state": "building",
   "subtasks": {
-    "completed": 0,
-    "total": 0,
+    "completed": 1,
+    "total": 14,
     "in_progress": 1,
     "failed": 0
   },
@@ -18,8 +18,8 @@
     "max": 1
   },
   "session": {
-    "number": 2,
+    "number": 3,
     "started_at": "2026-01-31T18:01:49.917579"
   },
-  "last_update": "2026-01-31T18:05:20.937003"
+  "last_update": "2026-01-31T18:07:18.842194"
 }
